Refactor latest main into the newer structure
