+++ src/qgis_geonode/ui/qgis_geonode_connection.ui
