+++ src/qgis_geonode/ui/qgis_geonode_layer_dialog.ui
