+++ src/qgis_geonode/ui/qgis_geonode_searched_resource.ui
